Fixed issue with disappearing item, added support for older Unity versions, etc
Hi, I really liked your approach to creating an inventory by separating the logic from the UI.  Thank you so much for this repository!

I decided to make a small fix to the problem when an item disappears when moved, when the origin movement cell is not equal to the origin coordinate of the item. I hope you'll approve the pull request I made.